搜索，百科模块测试完成

+++ digital-service/src/main/java/com/digital/event/EventConsumer.java
@@ -91,7 +91,7 @@ public class EventConsumer {
                 .isRead(0)
                 .build();
         Object targetUserId = commonEvent.getData().get("targetUserId");
-        message.setToId((Long) targetUserId);
+        message.setToId(Long.valueOf((Integer) targetUserId));
         if (Objects.equals(commonEvent.getTopic(), TopicConstant.TOPIC_COMMENT)) {
             message.setTopicId(TopicConstant.TOPIC_COMMENT_ID);
         } else if (Objects.equals(commonEvent.getTopic(), TopicConstant.TOPIC_LIKE)) {
@@ -108,7 +108,7 @@ public class EventConsumer {
         message.setContent(jsonString);
 
         // 系统发送提示消息
-        sseUtil.sendMessage((Long)targetUserId, String.valueOf(message.getId()), message.getContent());
+        sseUtil.sendMessage(Long.valueOf((Integer) targetUserId), String.valueOf(message.getId()), message.getContent());
 
         boolean save = messageService.save(message);
         if (!save) {
